Split ui into draw, coords and name resolver
Also add docs.
